pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
[tool.setuptools.package-dir]
world_map = ""
[tool.setuptools]
packages = ["world_map"]
[project.urls]
repository = "https://github.com/casperdcl/world-map"
[project]
name = "world_map"
version = "0.0.0"
authors = [{name = "Casper da Costa-Luis", email = "casper.dcl@physics.org"}]
description = "Poster-quality Eckert II projection of Earth"
readme = "README.md"
requires-python = ">=3.7"
keywords = ["svg", "map", "world", "poster", "cartography", "vector", "geospatial", "earth", "raster", "projection"]
license = {text = "MPL-2.0"}
dependencies = ["geopandas", "matplotlib", "networkx", "numpy", "pillow", "rasterio", "shapely", "svgwrite"]

[tool.flake8]
max_line_length = 120
exclude = [".git", "__pycache__", ".ipynb_checkpoints"]

[tool.yapf]
spaces_before_comment = [15, 20]
arithmetic_precedence_indication = true
allow_split_before_dict_value = false
coalesce_brackets = true
column_limit = 120
each_dict_entry_on_separate_line = false
space_between_ending_comma_and_closing_bracket = false
split_before_named_assigns = false
split_before_closing_bracket = false
blank_line_before_nested_class_or_def = false

[tool.isort]
line_length = 120
multi_line_output = 4
